refactor dry run into parachainBase
