+++ web/packages/api/src/toPolkadotSnowbridgeV2.ts
@@ -12,8 +12,6 @@ import { IGatewayV2__factory } from "@snowbridge/contract-types"
 import { OperationStatus } from "./status"
 import { FeeInfo, ValidationLog } from "./toPolkadot_v2"
 import { ApiPromise } from "@polkadot/api"
-import { Result } from "@polkadot/types"
-import { XcmDryRunApiError, XcmDryRunEffects } from "@polkadot/types/interfaces"
 import { accountToLocation } from "./xcmBuilder"
 import { Codec } from "@polkadot/types/types"
 export { ValidationKind } from "./toPolkadot_v2"
@@ -161,78 +159,6 @@ export function encodeNativeAsset(tokenAddress: string, amount: bigint) {
 // Encode assets array as bytes[] for the gateway contract
 export function encodeAssetsArray(encodedAssets: string[]) {
     return AbiCoder.defaultAbiCoder().encode(["bytes[]"], [encodedAssets])
-}
-
-export async function dryRunAssetHub(
-    assetHub: ApiPromise,
-    bridgeHubParaId: number,
-    destinationParaId: number,
-    xcm: any
-) {
-    const bridgeHubLocation = {
-        v4: { parents: 1, interior: { x1: [{ parachain: bridgeHubParaId }] } },
-    }
-
-    const result = await assetHub.call.dryRunApi.dryRunXcm<
-        Result<XcmDryRunEffects, XcmDryRunApiError>
-    >(bridgeHubLocation, xcm)
-
-    const resultHuman = result.toHuman() as any
-
-    const success = result.isOk && result.asOk.executionResult.isComplete
-    let forwardedDestination
-    if (!success) {
-        console.error("Error during dry run on asset hub:", xcm.toHuman(), result.toHuman())
-    } else {
-        forwardedDestination = result.asOk.forwardedXcms.find((x) => {
-            return (
-                x[0].isV4 &&
-                x[0].asV4.parents.toNumber() === 1 &&
-                x[0].asV4.interior.isX1 &&
-                x[0].asV4.interior.asX1[0].isParachain &&
-                x[0].asV4.interior.asX1[0].asParachain.toNumber() === destinationParaId
-            )
-        })
-        if (!forwardedDestination) {
-            forwardedDestination = result.asOk.forwardedXcms.find((x) => {
-                return (
-                    x[0].isV5 &&
-                    x[0].asV5.parents.toNumber() === 1 &&
-                    x[0].asV5.interior.isX1 &&
-                    x[0].asV5.interior.asX1[0].isParachain &&
-                    x[0].asV5.interior.asX1[0].asParachain.toNumber() === destinationParaId
-                )
-            })
-        }
-    }
-    return {
-        success,
-        errorMessage: resultHuman.Ok.executionResult.Incomplete?.error,
-        forwardedDestination,
-    }
-}
-
-export async function dryRunDestination(destination: ApiPromise, transfer: Transfer, xcm: any) {
-    const { registry } = transfer.input
-    const assetHubOrigin = {
-        v4: { parents: 1, interior: { x1: [{ parachain: registry.assetHubParaId }] } },
-    }
-    const result = await destination.call.dryRunApi.dryRunXcm<
-        Result<XcmDryRunEffects, XcmDryRunApiError>
-    >(assetHubOrigin, xcm)
-
-    const resultHuman = result.toHuman() as any
-
-    const success = result.isOk && result.asOk.executionResult.isComplete
-
-    if (!success) {
-        console.error("Error during dry run on source parachain:", xcm.toHuman(), result.toHuman())
-    }
-
-    return {
-        success,
-        errorMessage: resultHuman.Ok.executionResult.Incomplete?.error,
-    }
 }
 
 export async function getMessageReceipt(
