+++ web/packages/api/src/transfers/toPolkadot/erc20ToParachain.ts
@@ -11,8 +11,6 @@ import {
     claimerFromBeneficiary,
     claimerLocationToBytes,
     DeliveryFee,
-    dryRunAssetHub,
-    dryRunDestination,
     encodeNativeAsset,
     Transfer,
     ValidationKind,
@@ -404,6 +402,7 @@ export class ERC20ToParachain implements TransferInterface {
 
         // Check if asset can be received on asset hub (dry run)
         const ahParachain = registry.parachains[registry.assetHubParaId]
+        const assetHubImpl = await paraImplementation(assetHub)
         let dryRunAhSuccess, forwardedDestination, assetHubDryRunError
         if (!ahParachain.features.hasDryRunApi) {
             logs.push({
@@ -457,11 +456,10 @@ export class ERC20ToParachain implements TransferInterface {
                     "0x0000000000000000000000000000000000000000000000000000000000000000"
                 )
             }
-            let result = await dryRunAssetHub(
-                assetHub,
+            let result = await assetHubImpl.dryRunXcm(
                 registry.bridgeHubParaId,
-                destinationParaId,
-                xcm
+                xcm,
+                destinationParaId
             )
             dryRunAhSuccess = result.success
             assetHubDryRunError = result.errorMessage
@@ -475,7 +473,6 @@ export class ERC20ToParachain implements TransferInterface {
             }
         }
 
-        const assetHubImpl = await paraImplementation(assetHub)
         let destinationParachainDryRunError: string | undefined
         // Check if sovereign account balance for token is at 0 and that consumers is maxxed out.
         if (!ahAssetMetadata.isSufficient && !dryRunAhSuccess) {
@@ -530,8 +527,9 @@ export class ERC20ToParachain implements TransferInterface {
                                 "Dry run on Asset Hub did not produce an XCM to be forwarded to the destination parachain.",
                         })
                     }
+                    const destParachainImpl = await paraImplementation(destParachainApi)
                     const { success: dryRunDestinationSuccess, errorMessage: destMessage } =
-                        await dryRunDestination(destParachainApi, transfer, xcm[0])
+                        await destParachainImpl.dryRunXcm(registry.assetHubParaId, xcm[0])
                     if (!dryRunDestinationSuccess) {
                         logs.push({
                             kind: ValidationKind.Error,
